+++ client/scripts/views/pages/admin.ts
@@ -15,7 +15,7 @@ import { SubstrateAccount } from 'controllers/chain/substrate/account';
 
 import EdgewareFunctionPicker from 'views/components/edgeware_function_picker';
 import { DropdownFormField } from 'views/components/forms';
-import Tabs from 'views/components/widgets/tabs';
+import { Tabs } from 'client/scripts/views/components/component_kit/cw_tabs';
 import User from 'views/components/widgets/user';
 import { PageLoading } from 'views/pages/loading';
