+++ client/styles/components/component_kit/cw_tabs.scss
@@ -1,0 +1,27 @@
+@import 'shared';
+
+.Tabs {
+  .tab-bar {
+    border-bottom: 1px solid $neutral-100;
+    margin-bottom: 30px;
+
+    @include extraSmall {
+      margin-bottom: 16px;
+    }
+
+    .tab-entry {
+      text-decoration: none;
+      color: $neutral-600;
+      padding: 8px 10px;
+      margin-right: 4px;
+      display: inline-block;
+      font-weight: 500;
+
+      &.isActive {
+        color: $neutral-700;
+        border-bottom: 3px solid $neutral-500;
+        padding-bottom: 8px;
+      }
+    }
+  }
+}
