--- client/styles/components/widgets/tabs.scss
@@ -1,26 +1,0 @@
-@import 'client/styles/shared';
-
-.Tabs {
-  .tab-bar {
-    border-bottom: 1px solid $neutral-100;
-    margin-bottom: 30px;
-  }
-  a.tab-entry {
-    text-decoration: none;
-    color: $neutral-600;
-    padding: 8px 10px;
-    margin-right: 4px;
-    display: inline-block;
-    font-weight: 500;
-  }
-  a.tab-entry.active {
-    color: $neutral-700;
-    border-bottom: 3px solid $neutral-500;
-    padding-bottom: 8px;
-  }
-  @include extraSmall {
-    .tab-bar {
-      margin-bottom: 16px;
-    }
-  }
-}
